+++ dsmr_frontend/static/dsmr_frontend/js/dsmr-reader/live/voltage.js
@@ -15,6 +15,10 @@ $(document).ready(function () {
         tooltip: TOOLTIP_OPTIONS,
         calculable: true,
         grid: GRID_OPTIONS,
+        legend: {
+            top: '85%',
+            data: ['L1', 'L2', 'L3']
+        },
         xAxis: [
             {
                 type: 'category',
@@ -31,8 +35,9 @@ $(document).ready(function () {
                 min: 'dataMin',
                 max: 'dataMax',
                 axisLabel: {
-                    color: TEXTSTYLE_COLOR
-                }
+                    color: TEXTSTYLE_COLOR,
+                    formatter: '{value} V'
+               }
             }
         ],
         dataZoom: [
@@ -53,13 +58,15 @@ $(document).ready(function () {
         media: [
             {
               option: {
-                    toolbox: TOOLBOX_OPTIONS
+                    toolbox: TOOLBOX_OPTIONS,
+                    legend: {show: true}
                 },
             },
             {
                 query: { maxWidth: 500},
                 option: {
-                    toolbox: {show: false}
+                    toolbox: {show: false},
+                    legend: {show: false}
                 }
             }
         ]